% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_hsi.R
\name{extr_hsi}
\alias{extr_hsi}
\title{Title}
\usage{
extr_hsi(listlayer)
}
\arguments{
\item{listlayer}{the number of the his raster image layer, 1,2,3,4....}
}
\value{

}
\description{
Title
}
\examples{
lisddt <- 1:length(hsi@layers)
fdfyi <- extr_hsi(listlayer=lisddt )
saveRDS(fdfyi,'fdfyi.rds')
}
