% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrac_fam.R
\name{extra_fam}
\alias{extra_fam}
\title{my first fuction}
\usage{
extra_fam(lisr)
}
\arguments{
\item{lisr}{list of all months，data from multi_raslas_read function}
}
\value{
monall_fam
}
\description{
my first fuction
}
\examples{
matou_1 <- fread('E:/OneDrive - Business/闄跺闆<a8>/matou_data/matou1.csv')[,-1]
matou_1$num <-1: length(matou_1$type)
matou_2 <- readRDS('E:/OneDrive - Business/闄跺闆<a8>/matou_data/labelmatou2.rds')
matou_2$num <-1: length(matou_2$type)
####matou 1 raster
xy <- as.data.frame(matou_1[,c(7,6)])
xy$Lon <- xy$Lon-0.000006
xy$Lat <- xy$Lat+0.00001
xy <- as.matrix(xy)
e <- extent(xy)
r <- raster(e, ncol=500, nrow=500)
canopy1 <- rasterize(xy, r, matou_1[,12])
xy <- as.data.frame(matou_2[,c(7,6)])
xy$Lon <- xy$Lon+0.000005
xy$Lat <- xy$Lat-0.000002
xy <- as.matrix(xy)
e <- extent(xy)
r <- raster(e, ncol=500, nrow=500)
canopy2 <- rasterize(xy, r, matou_2[,12])
r_1to11 <- list(r2rmse_all1,r2rmse_all3,r2rmse_all4,r2rmse_all5,
                r2rmse_all6,r2rmse_all7,r2rmse_all8,r2rmse_all9,
                r2rmse_all10,r2rmse_all11)
 mon11_fam <- extra_fam(r_1to11 )
 factor(mon11_fam2$Fam)
 monall_fam <-mon11_fam\%>\%invoke(rbind,.)
 saveRDS(monall_fam,'monall_fam.rds')
}
