% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_raslas_read.R
\name{multi_raslas_read}
\alias{multi_raslas_read}
\title{to read and join the multi spectral with the cloud data}
\usage{
multi_raslas_read(las_list, dsf_list)
}
\arguments{
\item{las_list}{the list of each month cloud data with names}

\item{dsf_list}{the list of each month raster image data wihch form ras_band_ali}
}
\value{
a list of each tree in each month with multi spectra and cloud information
}
\description{
to read and join the multi spectral with the cloud data
}
\examples{
urll1 <- 'E:/PCGSPRO_1623035349/17366648332/42/map/'
dsf1 <- raster_read(urll1)
dsf1 <-dsf1 \%>\% unlist(recursive = F)\%>\%  unlist(recursive = F)
las_01 <- readLAS('E:/PCGSPRO_1623035349/17366648332/43/models/pc/0/terra_las/cloud.las',
filter = "-change_classification_from_to 1 2",
select = "xyzirc")
las_list <- list(las_01,las_03)
las_list <- list(las_01,las_03)
names(las_list) <- c('Jan', 'Mar')
dsf_list <- list(dsf1,dsf3)
names(dsf_list) <- c('Jan','Mar')
data_all <- multi_raslas_read(las_list,dsf_list)
}
