% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ras_ali_image.R
\name{ras_ali_image}
\alias{ras_ali_image}
\title{Title}
\usage{
ras_ali_image(monthi, fami)
}
\arguments{
\item{monthi}{the month number,for example: "May"}

\item{fami}{the family name of each tree,for example:"2_9_5"}
}
\value{

}
\description{
Title
}
\examples{
 data <- data
 fam <- (unique(data$Fam))
 month <- (unique(data$month))
 chl_tree <- ras_ali_image(fami=fam,monthi=month)
}
