% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ras_band_ali.R
\name{ras_band_ali}
\alias{ras_band_ali}
\title{Title to combine all singe multispectral image together fro DJI terra}
\usage{
ras_band_ali(url)
}
\arguments{
\item{url}{the link folder that contaim multiple singe layer images}
}
\value{
a list of all combined image data
}
\description{
Title to combine all singe multispectral image together fro DJI terra
}
\examples{
urll11 <- 'E:/PCGSPRO_1623035349/17366648332/117/map/'
dsf11 <- raster_read(urll11)
dsf1 <-dsf1 \%>\% unlist(recursive = F)\%>\%  unlist(recursive = F)
}
